consensus: test validator set change
